+++ Common/src/main/java/com/tristankechlo/additionalredstone/Constants.java
@@ -25,6 +25,7 @@ public class Constants {
         ModContainer.load();
         ModItems.load();
         ModItemGroups.load();
+        ModRecipes.load();
     }
 
     public static <T> T load(Class<T> clazz) {