+++ Common/src/main/java/com/tristankechlo/additionalredstone/blocks/CircuitBaseBlock.java
@@ -1,0 +1,23 @@
+package com.tristankechlo.additionalredstone.blocks;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.shapes.CollisionContext;
+import net.minecraft.world.phys.shapes.VoxelShape;
+
+public class CircuitBaseBlock extends Block {
+
+    private static final VoxelShape BASE = Block.box(0.0D, 0.0D, 0.0D, 16.0D, 2.0D, 16.0D);
+
+    public CircuitBaseBlock() {
+        super(Properties.copy(Blocks.REPEATER));
+    }
+
+    @Override
+    public VoxelShape getShape(BlockState $$0, BlockGetter $$1, BlockPos $$2, CollisionContext $$3) {
+        return BASE;
+    }
+}